Description: Use redundant u-boot environments
Implemented redundant environments in the bootloader. On boot it tries to load the environment at block A then block B, only if these both fail 
due to a CRC error is the default environment used. The saveenv command writes to the spare environment block before eraseing the know good 
block.

Also added a new command 'nandbb' that can be used to check for bad flash blocks.


Bug: 5152

git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/squeezeboxJive@507 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9