Bootloader changes from Ben Dooks.
1)
The attached patch alters the s3c2412 start code to check for the magic
number in INFORM0, and then start the resume process by using the address
stored in INFORM1. We have added a number of verification steps to this
to try and ensure that the resume process does not try and restart an
image that is not there, instead restarting uboot from scratch.

The verification methods include ensuring that the sleep mode was set
to power down, that the resume address lies within the SDRAM block
andthat there is a specific magic before the code to restart the kernel.

We have tested this code, leaving the unit sleeping overnight to ensure
that it will restart and continue working. There has been no power
consumption figures as there is no configuration for sleep mode pins.

2)
The ARM booting process passes arguments to the kernel which
include a machine identity number, which should be unique to
each board. Previously uboot has been passing an SMDK2413
identity through.

I have registered a new machine in the ARM Kernel
database at http://www.arm.linux.org.uk/developer/machines which is:
http://www.arm.linux.org.uk/developer/machines?action=list&id=1490

The attached uboot patch moves to using this new ID number. This
will affect booting 2.6.10 kernels, to continue using these with
this uboot you will need to add the new mach-types file and to
add an machine entry into arch/arm/mach-s3c2413/mach-smdk2413.c
for the new id.



git-svn-id: https://svn.slimdevices.com/repos/jive/trunk/squeezeboxJive@542 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9