Calculate revision number for SqueezeOS and SqueezePlay based on concatenation of  Squeezeplay and Poky  revisions.
Introduce a python method, squeezeos_squeezeplay_revision(), into squeezeplay_svn.bb.
This is used for both the SQUEEZEPLAY_REVISION in the squeezeplay build and for the SQUEEZEOS_REVISION when writing out /etc/squeezeos.version and the image files.

Calculate the revision number by using the stashed git revision count or subversion revision number for squeezeplay and appending the revision number or count for the poky build we are in. If both are using subversion then assume (require) they use the same repo and use only the one revision number.

The calculation is done in a new task, make_squeezeos_squeezeplay_revision, in squeezeplay_svn.bb. This stashes the result in staging for use by squeezeos_squeezeplay_revision().

squeezeos-image-boot.bb's do_rootfs task is made dependent upon squeezeplay's do_make_squeezeos_squeezeplay_revision task to ensure that the squeezeplay git revision count is available, although the install dependencies would in fact make this unnecessary.


git-svn-id: https://svn.slimdevices.com/repos/jive/7.8/trunk/squeezeos@9671 bbe22326-0783-4b3a-ac2b-7ab96b24c8d9