Ignore errors from "bitbake -c clean" operations
For some unknown reason we sometimes get an error relating to opkg and status.tmp when running "bitbake -c clean" on a target. I have been unable to find the cause of this but it is clearly benign. In any case, these clean targets are just not critical so ignoring errors should not be an issue.

Change-Id: I9c57df7ca26d97db813b9cedfd53740731f02e47